split layout for login/register page
