+++ src/views/LoginPage/LoginPage.css
@@ -1,80 +1,3 @@
-.login-page {
-  width: 97%;
-  height: auto;
-  display: flex;
-  justify-content: center;
-  padding: 145px 0 175px 2.5%;
-}
-
-.login-page-back-img {
-  position: relative;
-}
-
-.login-page-back-img img {
-  position: absolute;
-  left: -200px;
-  top: -155px;
-}
-
-.login-page-form {
-  z-index: 10;
-}
-.login-page-wrapper {
-  width: 400px;
-  height: 550px;
-  /* width: 50%;
-  height: 30%; */
-  background-color: white;
-  border: 2px solid black;
-  z-index: 1;
-  box-shadow: 7px 8px 0 0 rgb(0, 0, 0);
-  position: relative;
-}
-
-.login-page-cloud-img {
-  position: relative;
-  z-index: 20;
-}
-
-.login-page-cloud-img img {
-  position: absolute;
-  top: 490px;
-  right: -65px;
-}
-
-/* 로그인창 신호등 부분 */
-
-.login-page-action-bar {
-  border-bottom: 2px solid black;
-}
-
-.login-page-three-color {
-  width: 100%;
-  background-color: white;
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  box-sizing: border-box;
-  padding: 3px;
-}
-.login-page-color-circle {
-  width: 9px;
-  height: 9px;
-  background-color: #000;
-  margin-right: 1%;
-  border-radius: 50%;
-  border: 1px black solid;
-}
-.color-green {
-  background-color: #34b2a7;
-}
-.color-yellow {
-  background-color: #ffd43a;
-}
-.color-red {
-  background-color: #ff8473;
-}
-
 /* 로그인창 내용물 */
 .login-page-content-wrapper {
   display: flex;
