+++ src/views/RegisterPage/RegisterPage.css
@@ -1,19 +1,77 @@
+/* 로그인창 내용물 */
+.register-page-content-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 17%;
+}
+
+.register-page-input-wrapper {
+  display: flex;
+  flex-direction: column;
+  width: 80%;
+  /* margin-top: 20px; */
+}
+
+.register-page-icon-wrapper {
+  text-align: center;
+}
+
+.register-page-btn-wrapper {
+  display: flex;
+  margin: 10px 10% 0 0;
+  justify-content: flex-end;
+}
+
+.register-page-icon-wrapper img {
+  width: 5.7rem;
+}
+
+.register-page-icon-wrapper h2 {
+  font-size: 2.1rem;
+  margin: 10px 0 30px 0;
+}
+
+.register-page-input-wrapper input {
+  padding: 13px 20px;
+  margin-bottom: 5px;
+  border-radius: 40px;
+  border: 1px solid black;
+}
+
+.register-page-btn-wrapper button {
+  background-color: white;
+  padding: 10px;
+  border: 1px solid black;
+  margin-left: 5px;
+  box-shadow: 2px 2px 0 0 rgb(0, 0, 0);
+  cursor: pointer;
+  transition: 0.2s;
+}
+
+.register-page-btn-wrapper button:focus {
+  box-shadow: none;
+  transform: translateX(2px);
+}
+
 .register-page-nickname-wrapper {
   display: flex;
+  justify-content: space-between;
+}
+
+#nickname {
+  width: 61.5%;
 }
 
 .check-nickname-btn {
-  padding: auto 12px;
+  height: 44px;
+  padding: 0px 14px;
   background-color: #000;
   color: #fff;
   border: none;
   cursor: pointer;
   border-radius: 50px;
-  /* align-items: center; */
   box-sizing: border-box;
-  /* margin: 0 14px; */
   font-size: 14px;
-  /* font-weight: bold; */
-  /* white-space: nowrap; */
   justify-content: space-between;
 }
